remove precise value hover info and icon on devices that are touch screen. create displayStatistic.css and move it and DisplayStatistic.jsx into folder DisplayStatistic
